WIP - Add execution using next action
Missing:
- Track action to not execute twice
- Display execution progress in UI
